# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  models.AddedBooks:
    properties:
      book_count:
        type: integer
      week_date:
        type: string
    type: object
...
  models.BooksForList:
    properties:
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      like_count:
        type: integer
      sub_category_id:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
  models.CategoryBookCount:
    properties:
      book_count:
        type: integer
      category_name:
        type: string
    type: object
...
  models.DefaultResponse:
    properties:
      error_code:
        type: integer
      error_message:
        type: string
    type: object
...
  models.Meta:
    properties:
      current_page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
  models.StatisticCount:
    properties:
      author:
        type: string
      author_count:
        type: integer
      book:
        type: string
      book_count:
        type: integer
      top_book:
        type: string
      top_book_count:
        type: integer
      user:
        type: string
      user_count:
        type: integer
    type: object
  models.SubCategoryReq:
    properties:
      category_id:
        type: integer
      subcategory_name:
        type: string
    type: object
  models.SubCategoryRes:
    properties:
      _meta:
        $ref: '#/definitions/models.Meta'
      books:
        items:
          $ref: '#/definitions/models.BooksForList'
        type: array
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      subcategory_name:
        type: string
      updated_at:
        type: string
    type: object
  models.SubCategoryUpdate:
    properties:
      id:
        type: integer
      subcategory_name:
        type: string
    type: object
  models.Success:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Book store API
  version: "1.0"
paths:
...
  /statistic:
    get:
      consumes:
      - application/json
      description: Here get by parametrs count.
      operationId: getcounts
      produces:
      - application/json
      responses:
        "200":
          description: parametrs  Count
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StatisticCount'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: get parametrs count
      tags:
      - Statistic
  /statistic/category/bookcount:
    get:
      consumes:
      - application/json
      description: Here get by get category book count.
      operationId: getcategorybookcount
      produces:
      - application/json
      responses:
        "200":
          description: category book  Count
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CategoryBookCount'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: get category book count
      tags:
      - Statistic
  /statistic/week/bookcount:
    get:
      consumes:
      - application/json
      description: Here get by get week book count.
      operationId: getweekbookcount
      produces:
      - application/json
      responses:
        "200":
          description: Week book  count
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AddedBooks'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: get week book count
      tags:
      - Statistic
...
  /subcategory:
    post:
      consumes:
      - application/json
      description: Here category will be created.
      parameters:
      - description: Create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SubCategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubCategoryRes'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.DefaultResponse'
      security:
      - BearerAuth: []
      summary: create subcategory
      tags:
      - Subcategory
    put:
      consumes:
      - application/json
      description: Update Subcategory
      operationId: Subcategory
      parameters:
      - description: SubCategoryUpdate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SubCategoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Subcategory data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubCategoryRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Subcategory
      tags:
      - Subcategory
  /subcategory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subcategory
      operationId: deletesubcategory
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Success'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete subcategory
      tags:
      - Subcategory
    get:
      consumes:
      - application/json
      description: Here get by subcategory.
      operationId: getsubcategory
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: LIMIT
        in: query
        name: limit
        type: integer
      - description: PAGE_NUMBER
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: subcategory data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SubCategoryRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: get subcategory
      tags:
      - Subcategory
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"